src/main/java/ch/alpine/tensor/mat/NegativeDefiniteMatrixQ.java
// code by jph
package ch.alpine.tensor.mat;

import ch.alpine.tensor.Tensor;
import ch.alpine.tensor.TensorRuntimeException;
import ch.alpine.tensor.sca.Chop;
import ch.alpine.tensor.sca.Sign;

/** inspired by
 * <a href="https://reference.wolfram.com/language/ref/NegativeDefiniteMatrixQ.html">NegativeDefiniteMatrixQ</a>
 * 
 * @see PositiveDefiniteMatrixQ */
public enum NegativeDefiniteMatrixQ {
  ;
  /** @param tensor
   * @param chop
   * @return true if tensor is a negative definite matrix
   * @throws TensorRuntimeException if result cannot be established */
  public static boolean ofHermitian(Tensor tensor, Chop chop) {
    return StaticHelper.definite(tensor, chop, Sign::isNegative);
  }

  /** @param tensor
   * @return true if tensor is a negative definite matrix
   * @throws TensorRuntimeException if result cannot be established */
  public static boolean ofHermitian(Tensor tensor) {
    return ofHermitian(tensor, Tolerance.CHOP);
  }
}
